[rust] Option to skip empty fields
The properties buffer is a concatenation of:

[column_idx][column_data]

The length of column_data depends on the type of the column.  The type
of the column, in turn, is ascertained by looking up the column meta via
column_idx.

For numeric types (u8, i32, etc.), column_data is a fixed length (e.g.
for a u32 column, column_data is 4 bytes long), but for variable types
(String, Byte, Date, Json), the column_data is prefixed with a u32 len.

So altogether, the layout of a variable length field is:

[column_idx (2 bytes)][len (4 bytes)][data (len bytes)]

So that means it takes 6 bytes to encode an empty field, which can
really add up for large collections with sparse fields.

I'm assuming the reason each property in the properties buffer is
prefixed with its idx is in pursuit of supporting sparse properties —
that is, we can omit certain fields entirely.

If we don't want to support sparse properties, there's not much reason
to prefix each field with it's index (It also allows you to have a
random ordering of properties, but that doesn't strike me as
particularly important - is it?).

I'd like to take advantage of the format's support for sparse fields by
exposing an option to omit these empty fields from features.

Because I'm a coward, for now, I've made this an option and disabled it
by default so that there are not changes to existing behavior. But what
do you all think?